flask-admin-master/flask-admin-master/requirements-dev.txt
flake8
Flask>=0.7
Flask-SQLAlchemy>=0.15
peewee
wtf-peewee
mongoengine<=0.21.0
pymongo
flask-mongoengine==0.8.2
pillow>=3.3.2
Babel<=2.9.1
flask-babelex
shapely==1.5.9
geoalchemy2
psycopg2
pytest
pytest-cov
coveralls
pylint
sqlalchemy-citext
sqlalchemy-utils>=0.36.6
azure-storage-blob
arrow<0.14.0
colour
email-validator
wtforms==2.3.3
